navbar and blog
added blog page, which uses blogger api to load the pages. still under construction. still trying stuff out. william we need you
some of the titles and stuff are hardcoded rn and I would like those to be linked in the future

also I updated the navbar on all of the subpages so now it links to a separate file so if you update the navbar there (eg adding a tab or chaging a destination) it'll update for all of the pages. cool stuff. I need a life.