blog and changes
ok so now the blog page is pretty much fully functional. and the blogs load on the front page too so yay. it's all dynamically loaded, no hard coding woohoo. the navbar on the blog page still needs some work though, are we going to have two navbars or just one?